+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -2476,6 +2476,12 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
       }
 
       @Test
+      @TestMetadata("DaggerInteropDeeplyNestedClassNameExceedsFileLengthLimit.kt")
+      public void testDaggerInteropDeeplyNestedClassNameExceedsFileLengthLimit() {
+        run("DaggerInteropDeeplyNestedClassNameExceedsFileLengthLimit.kt");
+      }
+
+      @Test
       @TestMetadata("GenericMemberInjection.kt")
       public void testGenericMemberInjection() {
         run("GenericMemberInjection.kt");