+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsBoxTestGenerated.java
@@ -2166,6 +2166,12 @@ public class JsBoxTestGenerated extends AbstractJsBoxTest {
       }
 
       @Test
+      @TestMetadata("DaggerInteropDeeplyNestedClassNameExceedsFileLengthLimit.kt")
+      public void testDaggerInteropDeeplyNestedClassNameExceedsFileLengthLimit() {
+        run("DaggerInteropDeeplyNestedClassNameExceedsFileLengthLimit.kt");
+      }
+
+      @Test
       @TestMetadata("GenericMemberInjection.kt")
       public void testGenericMemberInjection() {
         run("GenericMemberInjection.kt");